Node configuration

Run configuration

Model architecture

MAGE.jl

Functions Bundles, Libraries and MetaLibraries

Bundles

Library

MetaLibrary

Casters

Pre made libraries

Element Operations

Pick Element

Integer Operations

Basic operations

Find First in vector

Modulo Operations

List Generic Operations

Basic operations

Subset functions

Make Lists from elements

Concat Operation

LIST Integer Operations

Basic operations

Comparison against 0

List Generic Operations

Basic operations

Arithmetic

Algebraic

Recursive functions

List String Operations

Capitalize Elements

Split String

Combinatorics

Integer Operations

Basic operations

Reduce functions

Reduce functions

String Operations

Basic operations

Grep Module

Paste

Conditional

Lower, Upper, Capitalize

Mutation Utils

Private Methods

Public Methods

Standard Mutation

Forced Mutation

Numbered Mutation

Decreasing Connection Mutation

Symbolic Regression

Node configuration

UTCGP.nodeConfig — Type

Specifies chromosomes/nodes properties

n_nodes::Int(
    connection_temperature::Int
    arity::Int
    offset_by::Int)

Run configuration

UTCGP.runConf — Type

runConf(lambda_::Int
    generations::Int
    mutation_rate::Float64
    output_mutation_rate::Float64)

Specifies the experiment properties.

Model architecture

UTCGP.modelArchitecture — Type

Specifies the model types for :

inputs

program

outputs

Outputs types should be a subset of program (chromosome) types.

In principle, input types could be different than program/outputs types although that case is rare

modelArchitecture(
    inputs_types::Vector
    inputs_types_idx::Vector{Int}
    chromosomes_types::Vector{<:DataType}
    outputs_types::Vector
    outputs_types_idx::Vector{Int})